dummy_microservices/oauth/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
pydantic
pytest 